// src/generated/api/models/CommentDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserSummaryDto } from './UserSummaryDto';
import {
    UserSummaryDtoFromJSON,
    UserSummaryDtoFromJSONTyped,
    UserSummaryDtoToJSON,
    UserSummaryDtoToJSONTyped,
} from './UserSummaryDto';

/**
 * 
 * @export
 * @interface CommentDto
 */
export interface CommentDto {
    /**
     * 
     * @type {number}
     * @memberof CommentDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof CommentDto
     */
    content?: string;
    /**
     * 
     * @type {boolean}
     * @memberof CommentDto
     */
    active?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof CommentDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CommentDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {UserSummaryDto}
     * @memberof CommentDto
     */
    author?: UserSummaryDto;
    /**
     * 
     * @type {number}
     * @memberof CommentDto
     */
    postId?: number;
}

/**
 * Check if a given object implements the CommentDto interface.
 */
export function instanceOfCommentDto(value: object): value is CommentDto {
    return true;
}

export function CommentDtoFromJSON(json: any): CommentDto {
    return CommentDtoFromJSONTyped(json, false);
}

export function CommentDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommentDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'content': json['content'] == null ? undefined : json['content'],
        'active': json['active'] == null ? undefined : json['active'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
        'author': json['author'] == null ? undefined : UserSummaryDtoFromJSON(json['author']),
        'postId': json['postId'] == null ? undefined : json['postId'],
    };
}

export function CommentDtoToJSON(json: any): CommentDto {
    return CommentDtoToJSONTyped(json, false);
}

export function CommentDtoToJSONTyped(value?: CommentDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'content': value['content'],
        'active': value['active'],
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'] == null ? value['updatedAt'] : value['updatedAt'].toISOString(),
        'author': UserSummaryDtoToJSON(value['author']),
        'postId': value['postId'],
    };
}